+++ gendarme/rules/Gendarme.Rules.Smells/AvoidSpeculativeGeneralityRule.cs
@@ -226,7 +226,8 @@ namespace Gendarme.Rules.Smells {
 
 		public RuleResult CheckType (TypeDefinition type)
 		{
-			if (type.IsEnum || type.IsGeneratedCode ())
+			if (type.IsEnum || type.IsGeneratedCode () ||
+                type.Name.Contains("@"))
 				return RuleResult.DoesNotApply;
 
 			CheckAbstractClassWithoutResponsability (type);