+++ gendarme/rules/Gendarme.Rules.Maintainability/Test/AvoidUnnecessarySpecializationTest.cs
@@ -661,5 +661,13 @@ namespace Test.Rules.Maintainability {
 		{
 			AssertRuleSuccess<TestCase690057> ("WyslijMaile");
 		}
+
+        [Test]
+        public void FSharpIgnoreClosures()
+        {
+            var probe = typeof(AvoidMultidimensionalIndexer.DotNet.CLIArgs);
+            var type = probe.Assembly.GetType(" AvoidUnnecessarySpecialization.Main+prepareTargetFiles@22-3");
+            AssertRuleDoesNotApply(type, "Invoke");
+        }
 	}
 }