+++ ReleaseNotes.md
@@ -1,4 +1,4 @@
-# 2020.2.xx.xxxxx-pre-release
+# 2020.2.17.170xx-pre-release
 F# niggles part 4
 * Exempt generated types with `@` in their names from `AvoidUnnecessarySpecializationRule`, `AvoidSpeculativeGeneralityRule` and `MethodCanBeMadeStaticRule`
 * Exempt F# placeholder arguments `_` (compiled to `_arg...`) from `UseCorrectCasingRule`